dataflow: Wrap domain execution in a span
I added this logging to help debug an issue, but it seems generally
useful - thesse two functions seem like the primary entrypoints to
handling within a single Domain, so wrap each in an `info_span!` so we
can tell which log events are coming from which domain. Note that this
does *not* use the `tracing::instrument` proc macro since that's limited
to only using the Debug impl of the fields, and I want to use the
Display impl of the domain index.

Change-Id: I5e4637b3acc99105832529a561e3c1f344ec255d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/614
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>