packet_filter: Use inclusive naming, clean up formatting
The word "whitelist" is considered "Tier 1: Replace Immediately" by the
Inclusive Naming Initiative [0] - not to mention that I found it a
little unintuitive as used in this module. This commit renames:

* WhitelistData to NodeKeys, since it contains the set of keys requested
  by a particular node
* PacketFilter::whitelist to PacketFilter::requested_keys, since it
  contains the keys requested by all child nodes

And also updates all of the places in comments where we say "whitelist"
to say "allowlist" instead.

Also, since I'm updating documentation anyway, this updates the
formatting of the comments in this module to properly wrap to 100
chars (and no shorter, like in certain places), as consistent
line-wrapping of comments greatly aids readability.

[0]: https://inclusivenaming.org/word-lists/tier-1/

Change-Id: Ifa804664124827bcea8de08a54b4af26f42a2cbb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/615
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI