packet_filter: Fix boolean logic in filtering
Previously the packet filter's filter check for records would
incorrectly allow through only records that matched *all* previously
requested keys, rather than *any* of the previously requested keys. The
result of this, effectively, was that if we ever replayed more than one
distinct key to a reader all subsequent updates would get dropped
entirely (!).

This commit fixes that, refactoring the logic to use `Iterator::any()`
rather than nested for loops and a short-circuiting return, which makes
the behavior much more obvious. Included are a unit test in packet
filter and a noria-server integration test for the original issue.

Change-Id: Idfb079a25d5737f704d63b9a59113a2f939a2d9c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/616
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>