Order/Limit Removal: Support additional syntax
Prior to this change, the order_limit_removal pass only looked for keys
defined in the constraints section of the column definitions when
checking to see whether a column represented a unique or primary key.

This change checks the keys field in the table to see if the column is
listed as a unique or primary key. This change also removes checking the
column constraints for the primary key constraint, since the
key_def_coalition pass will add primary keys to the table.keys field.

Compound primary and unique keys are ignored.

Change-Id: I53c36797537a759add404ef100db515256e81f4b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/617
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>