server: Disable packet filtering by default
Disable the creation of packet filters for egresses before readers by
default, as it's an experimental and incomplete feature that's not
indended to go to production, and add a new --enable-packet-filters flag
to noria-server to enable it again.

Since the integration_serial::it_works_basic test exercises the metrics
reported by packet filtering, this enables packet filtering in those
tests explicitly.

Change-Id: I4cea64e8a2d6dc16e905cf1d5bc8ec01622bc6aa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/619
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>