readyset-substrate: Comment plan in gerrit
When we finish the terraform plan for a particular root module, generate
a description of that plan and post it to Gerrit as a "robot comment" on
the patchset, along with a link to the buildkite step that generated the
plan.

Under the hood, we're now passing around a new terraform::Plan struct,
which saves a reference to a NamedTemporaryFile containing the
serialized plan, knows about the module path the plan was created for
and whether there were any changes, and has methods to shell out to
terraform to get a description of that plan, either human-readable or as
JSON. This is used both when uploading the plan (rather than just
assuming the plan file is going to be in a predictable location) and
when commenting on Gerrit.

Change-Id: I99af6ae0f26ec889a853e3fe1bff8436ca67769b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/622
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>