passes: Don't always negate a bare IN
Previously the negation_removal pass was always incorrectly negating a
bare IN condition, regardless of whether it was surrounded in a NOT -
this updates that to only happen if the `negated` flag passed in is
`true` (because it was surrounded by a unary NOT) and covers it with a
couple of unit tests.

This also includes trace-level logging of the query before and after
rewriting, which I added to help with debugging this issue but seems
like it'd be generally useful.

Change-Id: I99b978c5562d480671171c340db8507f28497153
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/627
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>