noria-client: Don't rewrite constant-valued IN
Previously, I *assume* to work around a lack of support for
constant-valued in in noria-server, the *client* rewrote IN in filters
to be a parametrized multi-lookup, with the parameters stored alongside
the parsed query in the parsed_query_cache and passed around as
`use_params` everywhere the query was used. This extra complexity is
totally unnecessary now that we natively support IN in noria-server, so
let's get rid of it.

Change-Id: Iab2eb68c41b841e0e12be065c3b36578c7c92d9d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/629
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>