Wait for replication snapshotting in noria-*sql tests.
It is possible that queries are not ready to execute in noria
while replication snapshotting is occuring. However, the
check for query rediness used in many tests will indicate
that it is ready. This fixes that behavior by adding an async
notification when the snapshotting completes.

This is implemented as a notification the leader should satisfy
when it is created and fully ready.

The noria-*sql tests now verify query readiness before
starting.

Change-Id: If1d48543edc57bbdfebb924fd72f8811252bcaee
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/633
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>