Bound the exponential backoff in reader
Currently when a reader requires an upquery it will recheck if the upquery
is finished every given period of time. In addition when the recheck fails
for too long (usually because the values were evicted between the time the
upquery was issued, and the retry was performed) it will reissue the upquery.

This process repeats with an exponential backoff. If evictions happen a lot
this leads to an unbounded backoff growth, where it even exceeds the view
timeout.

This change attempts to put an upper limit to this exponential growth.

Change-Id: I5bed36a8d88b3c1a2ccc6b28347d29978909d569
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/638
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>