ops: Add terraform module for ops interview
Add a new Terraform module for provisioning all the AWS resources needed
to conduct our current "ops fizzbuzz" interview. This is parametrized on
the candidate's name, public IP address, and SSH key (which we ask for
before the interview) and the interviewer's IP (if necessary) and ssh
key (to make a key pair for the instance), and provisions a brand new
isolated VPC, and security group for the interview, ready to be torn
down after the interview is done.

Change-Id: I6b37a702f4de0191f5b1803a5e5f73a258f99296
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/639
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI
Terraform-Plan-Review: Harley Klein <harley@readyset.io>