noria-client: Disallow parametrized IN
IN with parameters is currently implemented via a multi-key lookup to
the view with the results concatenated - this is fine for normal
queries, but for aggregates all filters - even parametrized ones - are
supposed to happen *before* the aggregate. The end result here is that
IN with parameters ends up with one row per parameter, rather than a
single globally-aggregated row.

We *could* hypothetically support this (except for AVG) by aggregating
post-read, but that's not in scope for right now - for the time being,
we should just return an unsupported error for these queries so they go
to fallback. This does that, and adds a new fallback integration test
checking that the query overall returns the correct results

Fixes: ENG-624
Change-Id: I587948358990fd00c583048e4bae5f463fe24533
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/641
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>