Don't log at > info in release builds
Add the `release_max_level_info` feature flag to all places in the tree
we depend on the tracing crate, so that release builds don't even have
tracing logging compiled-in.

This is not just for performance (though that's not irrelevant here) but
also so that we can log internal-only information or data in queries or
resultsets at will at levels debug and trace without being worried that
a customer will see the former or accidentally log the latter by
enabling a particular log level in production.

Since cargo merges feature flags for all instances of a given dependency
in a workspace, this probably could've been done by just adding the
feature flag to the tracing dependency in readyset-logging, but I wanted
to be extra careful here.

Change-Id: I49b0b64f8ed0e136177289dc4425a9080f1f09e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/642
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>