Added support for JSON and JSONB types for Postgres.
We reuse the `SqlType::Json`, with now a correct way of encoding/decoding the
json value in Postgres.
The `serde_json` library was introduced only for parsing/generating valid
`serde_json::Value`s.

This commit also does not implement `Literal` JSON value parsing for Postgres,
like `'[1, 2, "foo", null]'::json`.

There's a caveat here regarding `JSONB` and the way we are testing it:
For starters, we are using `serde_json` as the library that actually handles the
JSON values, parsing, which `tokio-postgres` extends by adding Postgres serdes
(`FromSql`/`ToSql`).
From `serde_json` we use the `arbitrary_precision` feature, which stores Json
numbers as strings to ensure the precision is not lost (and we use this to be
compliant with PostgreSQL Json numbers, which are `NUMERIC` behind the scenes).
The problem is that there are slight differences between the values picked by
the WAL reader and the values generated with our proptests in terms of JSON
numbers differ syntactically (not semantically), causing the tests to fail.
For example, some failing cases are comparisons like `0.0 == -0.0` or `0.000054
== 5.4e-6`, which fail because of a literal check.
A potential fix to this would be to modify our `tokio-postgres` library and
ensure that we use `arbitrary-precision` in its `serde_json`, since maybe the
serdes is not using it, so there's a difference in representation.
The best fix would be to fork/contribute to the `serde_json` library to allow
using `rust_numeric::Decimal` types as its `Value::Number` representation.

Refs: ENG-604
Change-Id: I02df21e4cca672cd43b9b5f6535c0117bb22c4ed
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/643
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI