Add a way to link prepared statements for qca
When QCA is enabled we need a way to link prepared statements together.
This is because we always run all queries against both Noria and the
upstream database, and always return the results from the upstream
database.

Because of this, when we get prepares, we will have returned the
statement id for the upstream prepare. So when the client uses that
statement id to issue an execute, we will only execute against the
upstream database.

By providing a way to link the two ids together, we can properly ensure
that we are running executes against both Noria and our configured
upstream database.

Change-Id: Icc7a6bdf412c62369824abc1b03b4b4f5d74ca71
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/647
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>