Add mirror_execute and clean up execute
This commit adds a mirror_execute method that operates similarly to
mirror_read and mirror_prepare helper methods. Using it ensures that we
run an execute against both Noria and our upstream database. This only
works if QCA is enabled. In those cases we would have run prepares
against both upstream and Noria. This means we have a moment where we
can link the two statement ids together.

This is done in a way that's no-op if QCA is not enabled, because the
hashmap that is used to link the two ids lives on the query coverage
info struct.

When we get to mirroring the executes then, we simply need to retrieve
the noria statement id that is linked to the upstream statement id the
client would have given us. Then we simply execute against both.

This commit also refactors execute to be a bit cleaner with separate
methods for executing against noria vs executing against upstream.

Change-Id: Id5789792fffe10fe8a2b6ef7f231cb9b75a3bac3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/649
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>