Remove slowlog for executes
It was deemed in a slack discussion that we should remove slowlog
entirely. Slowlog is unused and superflous in the presence of QCA. This
commit removes one use of it that's currently blocking a refactor I need
to do in facilitate mirroring executes in the case that QCA is enabled.
Leaving slow log would involve unnecessary earlier
cloning.

Change-Id: Id5015db36a8c88305bdcc7206c40cd2d3f2e5a8c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/648
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>