readyset-substrate: handle SIGTERM better
Previously, if `terraform plan` run running and the runner process
recieved SIGTERM, the parent process would end and the state might be
left locked by the process when the host was collected to run another
agent.

This change forwards the SIGTERM signal that was sent to the process to
the subprocess so cleanup can begin.

Supporting this first on just `terraform plan` since that is what is
currently running in CI. Making this reusable and wrapping more commands
with it is left for a future change.

Change-Id: I715236b98a55ed66f58cee71fbde33b5ed5476a5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/652
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>