Don't allocate ports on the host machine for CI docker containers.
docker-compose.yaml currently allocates ports on the host machine
for each of the docker containers. This can be a problem when
docker containers on the host are run in parallel, which happens
when multiple steps of a buildkite pipeline are run in parallel.

As a result, we occasionally get "port already allocated" failures
in CI. This updates the base docker-compose.yaml to use `expose`
instead of `ports`, exposing ports on on the docker-compose's
internal network but not on the host machine.

Change-Id: I5a04813f8c6c2b7556b6007c6262bee96aa9dc6a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/653
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>