Added support for `BIT` and `VARBIT` types for Postgres.
A new `DataType::BitVector` was added, which holds a `BitVec` from the `bit-vec`
library.
Couple of notes to make here:
1. The `nom-sql` support to parse a literal bit vector is available through the
`Dialect`, which yields a `BitVec` for PostgreSQL and fails for MySQL (since
MySQL doesn't support bit vectors).
2. We are displaying the `BitVec` as a string of 0s and 1s, per PostgreSQL
specification.
3. While it's true that `ps::Value::Bit` and `ps::Value::VarBit` are identical
in terms of representation and serdes, I added the distinction to be able to
correctly match those to their PostgreSQL types (`Type::BIT` and `Type::VARBIT`
respectively).
4. I just implemented the basic coercion between `Bit` and `VarBit` SQL types.
I am not sure if bit vectors coerce to other types (probably they do), but
implementing them is out-of-scope for this task.

Refs: ENG-607
Change-Id: Ie06246c0c08555383f434deb94af6a17454be137
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/654
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>