readyset-substrate: Don't lock when planning in CI
Currently, we're running into a recurring issue in CI where if the
terraform plan steps are cancelled (eg because the build is
automatically cancelled due to another patchset being pushed to the same
CL), we leave the state lock dangling, causing all subsequent builds to
fail without any manual intervention. We're not actually *applying* the
plans we're making in terraform yet, so in lieu of trying to install
some custom signal handlers to exit gracefully I think it makes the most
sense to just pass -lock=false to the terraform plan invocation so we
don't create the locks in the first place.

Change-Id: I33bda559950aeca1df1159225cc8fdf46f5e00d7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/655
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>