ops: Fixes from cfn testing.
Need to use a managed policy to be able to get an ARN which can be
shared between multiple stacks.

Made sure to pass Consul tags through correctly to adapter and server
launch configurations.

Fixed an issue where the password to the database wasn't getting
subsituted which was causing user-data to not be run.

The name of the unauthenticated connection parameter changed for the
adapter.

Allow gossip for Consul ports 8300 and 8301.

Change-Id: Ie8e091b5eafdf9289627d79eb791f66879cf24d7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/668
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>