noria-client: Pass a SelectStatement to prepare_select
Everywhere we call prepare_select we have a SelectStatement, then we
wrap it in a `SqlQuery::Select` - only to pattern match on it again in
`prepare_select` and return an internal error if it's anything other
than a `SelectStatement`. This is pretty silly - let's just make it
accept only a SelectStatement.

Change-Id: I300f1de3ea0d3eebecf238875ebe8350844e0fcf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/671
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>