noria-logictest: Correct float parsing
The previous way we were parsing and storing floats in noria-logictest
was broken for a number of reasons - most notably, we were always
modifying the fractional part by a constant number, regardless of how
many trailing zeros it had. This essentially meant that any float other
than 1.0 would never match the results returned by the database. Instead
of trying to redo this in the parser, this commit updates this to just
convert to an f64 and then use `Value as From<f64>` to get out the
Value.

Change-Id: Ie013a768c3e1760dc6cb6b4a3fb1f730d94c5de5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/673
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>