Accept DDL in prepares if mirror_reads is enabled
We should accept all DDL in prepares as long as mirror_reads is enabled.
This DDL will hit upstream. We will fail to replicate, but that's
perfectly fine because all future reads (including prepared reads) will
hit both, and we always return the result from upstream. That means that
an end user using QCA will always get the correct result even in the
face of dropped, altered, or renamed table DDL.

Change-Id: Icccece3984de98ee4ce03acd3ab9cc080abc1d45
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/681
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>