Deploy vector in readyset-server and readyset-mysql-adapter AMIs.
Vector will be used for log and metrics aggregation in our deployments.
This CL creates the framework for aggregation from each readyset-server
and readyset-mysql-adapter image. It collects basic logs from journald,
in the future this will be expanded to include node_exporter and
prometheus metrics.

The vector agent runs as a service on the readyset-server and
readyset-mysql images. On each of the images, /vector_agent
initialization and files sets up the config required for
the vector agent specificially, while /vector manages
the running the service with /vector_agent's config. This
is set up as a two stage process as in a future CL the
vector service can run with separate config used for log
aggregation.

Change-Id: Id2cb70b5ba8064d70784dec493e8ff5467144027
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/682
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>