Fix QCA gzip encoding error
In testing it appeared that we weren't correctly saving the QCA bundle.
Unpacking the bundle always resulted in unexpected EOF, and in partial
data from the extracted files.

This commit ensures that we do the necessary logic to correctly
finishing writing the bundle, and error log if something does fail in
the finalization process.

Tested and working.

Change-Id: I533bdbfe0dfdc5dbc58b6d18b7cab35651dd38b8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/689
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI