dataflow: Don't match on NULL in join keys
The semantics of NULL in SQL are tri-state - while obviously `1 = 1`, it
is *not* the case that `null = null`. Usually this is irrelevant for
lookups into state since it's impossible to upquery for null keys (since
IS and IS NOT can't be parametrized, syntactically), but - it turns out
- this *is* relevant for JOINs, which use regular state key queries to
do lookups into their parents as part of processing joins. Previously
those state lookups would match a NULL in a join key on one parent with
a NULL on the other parent, causing us to match on the join where a
regular SQL database wouldn't. Now, we just explicitly check as part of
the join processing code whether the input key has any NULL values in
it, and if so we short-circuit to return no rows for the other side of
the join, without doing any lookups.

This fixes 2 previously failing logictests, plus comes with an
integration test for the join operator.

Fixes: ENG-655
Change-Id: Ie43c0bac81d61881da38626e0f92ac81966b7505
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/690
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>