Remove LIMIT/ORDER when returning at most one result
This applies the remove_order_limit pass to queries in noria. This pass
removes the order and limit fields of a query if it can determine that
the query will return at most one result.

To return the correct prepare response to the client, the adapter parses
the order/limit clause in the query to see if there are any
placeholders. If so, it adds them to the parameter columns written back
to the client. It also stores them as ignored columns, so that the
corresponding parameters are dropped when sent by the client as part of
an execute statement. The adapter will check the values sent by the
client for limit and offset and short-circuit if they guarantee an empty
resultset.

The adapter always ignores LIMIT/ORDER placeholders, because they are
either removed by the server, or the server will fail to install the
query in noria.

This change also updates the order limit removal pass to handle table
aliases. The pass will not attempt to resolve aliases until it fails to
find a table with the column name. This assumption behind this ordering
is that most queries will not use table aliases, so it is better to have
a worse execution time for this path in the worse case for a better
execution time in the average case.

This implementation is a bit hacky, and a better implementation is
planned:
https://docs.google.com/document/d/1TOe9RGwtQ2eYQjvU4flprR_UA1p-zMN-6Tfajt45pPk

Change-Id: I305dd355f7718e04735add15ae4758e399966270
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/692
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>