dataflow: Explain why Extremum doesn't need to evict
There was a bit of a discussion that happened on ENG-356 around this,
where originally we had thought that evictions that lost state would
need to evict downstream of themselves similar to joins when they lost
that state, but after some thought we realized that that's not
necessary, since because of the rules of partial if an extremum misses
when looking up into its parent that means its descendants *can't* have
that key, so dropping writes as usual is totally fine to do. That only
isn't the case for joins because they have *two* parents, and might be
looking up into a parent that they haven't just received an update from.
This documents that realization in a comment in ops/grouped/mod.rs
where we miss while handling the GroupedStateLost error.

This is relevant right now mostly for topk, since my thinking is that
this same property will hold there, and allow a topk that's just dropped
below k records to do the same thing where it just looks up the records
in its parent.

Fixes: ENG-356
Change-Id: Ifb3fc80eafa60765f3b82333a85eed72c239b573
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/693
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>