Remove FallbackNoConnector Error
Instances of the FallbackNoConnector error are replaced with Internal
errors in this CL. These errors were placed in cases where we attempt to
execute a query in fallback, but do not have a fallback connector
configured.

However, the adapter is currently designed such that it should not
attempt to execute any queries over fallback if an upstream connector is
not configured. Therefore, an Internal error is more appropriate in
these instances.

Change-Id: Idc24755e3907fea1f87cd36ed3c3f94b199aa9aa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/696
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>