Set the CONSUL_BOOTSTRAP_EXPECTED env variable in the consul template.
The CONSUL_BOOTSTRAP_EXPECTED value should be set to the number of
consul servers required for a quorum in the deployed cluster.
Otherwise no quorum is achieved and each client acts as it's own
cluster. This manifests in not all writes and updates to KV entries
being propagated. As a result, the readyset adapter and readyset
server's may not be aware of who the current controller is.

Change-Id: I08317139e6c25e15219d9e7152f95c6f657b3a42
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/698
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>