Register all workers with new leader on leader failure.
When a new leader is elected, this leader should register
all workers currently in the system. This is typically done
through a `NewWorkers` message that is propagated when
the worker state is detected to be changed. Due to the way
we previously hadnled state, no `NewWorkers` messages were
propagated preventing any existing workers from entering the
system.

Added resetting the active workers when the leader election
state is indicates it is first elected a new worker and
moved the is_leader assignment to the correct place.

This is tested via clustertest.

Change-Id: I02db1d98eb4be103267e24eab74e2d2a93efcb4b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/706
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI