Reimplement the Text and TinyText to support UTF-8
Previously the `Text` type was implemented as an ArcCStr, but
that type does not support UTF-8 and is also very slow to properly
serialize and deserialize as it has to call the libc `strlen` function
every time it calls `as_str` or `as_bytes`. The new implementation
uses a ThinArc<[u8]> internally and it validates the contents are proper
UTF-8 at creation time. Afterwards all operations are straightforward
derefs, including `as_str` and `as_bytes`.

The serialization and deseriazlization are also much faster now, and
leads to notable improvements in `noria-mysql`.

The reason to using ThinArc<[u8]> as opposed to Arc<Vec<u8>> are twofold:
To avoid an extra indirection, to use less storage (as triomphe::ThinArc
is not storing a weak count, and a slice is not storing capacity like a
Vec does), lastly ThinArc is faster to drop than Arc.

`TinyText` was also reimplemented as a new struct. Since interior zeroes
are allowed in UTF-8, it includes a length field now. `as_str` is now also
a zero cost operation.

Change-Id: I3c362b1dc63ada8f22a309ca81bf51527668d6c3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/707
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>