Added support for `TIMESTAMP WITH TIME ZONE` type for Postgres.
Added a new `DataType::TimestampTz` that uses a `DateTime::<FixedOffset>` to
represent the `TIMESTAMP WITH TIME ZONE` type, to make use of the `ToSql` and
`FromSql` implementations offered by `tokio-postgres`.
We also added a new `SqlType::TimestampTz`.

As a side note, the `wal_reader.rs` code in charge of parsing the `DateTime` is
not ideal, because the WAL reader cannot represent negative years, and to handle
them it adds an epoch ("BC" at the end if the year is negative).
We are handling that scenario, but ideally we would want to add a parser in
the future.

Change-Id: Ibec38c3d4daf49795b3fa7e9d218f5e83f454ae2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/688
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI