noria-mysql: Implement MVP of vertical testing
This commit adds a new integration test suite to noria-mysql called
"vertical.rs", which provides an MVP of implementation of the "vertical
testing" design doc[0]. That doc has more information, but in short, the
idea here is that we're writing a suite of proptests with hand-selected
queries that generate a sequence of *operations* to perform on those
queries, then perform those operations on both mysql and noria and
verify that the results match. This commit, being just an MVP, only
implements the "Insert" and "Query" operators, and only defines one test
with one very simple query - but lays the groundwork for writing even
more tests. In more detail:

- When generating Query operations, we either generate a query with a
  brand new key, or we generate a query with a key that's *already been
  inserted*, which we do by using the set of operations that have
  already been generated every time we generate a new operation.
- There's a `vertical_tests!` macro, which allows concisely defining a
  query and the schema for the tables that query uses to run a test on.

Note that while the eventual plan is to skip this suite as part of the
normal build and have it run with a very high PROPTEST_CASES value in
the nightly build, for now it's passing in just a few seconds so it's
left running as part of the normal build.

[0]: https://bit.ly/3FwrvKo

Fixes: ENG-664
Change-Id: Id19831ff94aa65b1081c8804f92293b37706f4f5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/719
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>