ops/cfn: Add VPC endpoint
This adds an Endpoint Service which will allow the creation of an
Endpoint to be created inside any other VPC to connect to the NLB inside
the VPC being created by the stack.

This endpoint will not be managed by us and must be cleaned up before
the stack can be removed.

Change-Id: Ic499fc0f21e2f4578275388f27a9e3b6b33f8c96
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/713
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>