ops/cfn: Allow changing Consul tag. Test script.
Passing through the various Consul related attributes. This is needed
for the testing script also included which makes sure to not share
the tag across environments.

The script should be able to be run with just the ssh key name and get
a stack.

Ensure that a deployment name is passed through that is more than one
character.

Change-Id: If0789d5ed37547111b4ad2bd80f114346465e7bc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/714
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>