Update to json log format and include app identifier and host.
The existing log format only logged each server's log message at
the aggregator, without additional information about the source
of the message. Changing to the json format automatically gets us
all the information we need from the syslog appended fields, however,
this leads to a difficult to parse log message. Vector remapping
is performed to parse out the important parts of the message
we are interested in.

Change-Id: Icfbe433d1ec014b20b5ff560cc701dd3768d714e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/716
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>