Implement SelectSchema fields as Cows
Currently we clone the select schema every time we execute
a prepared statement, but that is an expensive operation as
the schema may indeed be larger than the resultset (as I
found out in my IRL benchmark). Reimplementing the SelectSchema
field as Cows improved performance by a big margin.

Sadly I had to move some code around in a sub optimal way to
convince borrowck that everything is kosher.

Change-Id: Ia06e66776cdd4c4c649f07fb4d8a7f14d4b99028
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/721
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>