treewide: Update dependency on nix (via mysql)
Update the dependency we transiently have on the `nix` crate to patch
RUSTSEC-2021-0119[0], primarily via updating our dependency on the
`mysql` crate. Initially, this was done by explicitly setting the
version of the mysql crate we depend on to 21.0, but even that
specifically only depends on nix version 0.21.x - there's a patched
version 0.21.2 in that series which doesn't have RUSTSEC-2012-0119, but
that conflicts with other crates in our workspace tree - rather than
trying to manually solve for a dependency that's going to work (hello,
np complete problem), I went with just making a (temporary) fork of
rust-mysql-simple with the `nix` dependency updated.

[0]: https://rustsec.org/advisories/RUSTSEC-2021-0119

Change-Id: Ia13303a61c156673ef6af5f2218809440dc099a0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/722
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI