server: Log when domains exit with an error
Domains can either panic, in which case polling the JoinHandle will
resolve to Err(JoinError), *or* they can return Errors, in which case
polling will resolve to Ok(Err(error)) - we handled the former case, but
the latter case was getting caught up in the catchall "domain exited
unexpectedly" code. This matches on the latter case explicitly to
actually log the error that caused the domain to exit, fixing up
indentation along the way.

Change-Id: Ibc24a53274d6a42c79530ab51c8dcf699b102ca9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/727
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>