topk: Support partial deletes by querying parent
Implement proper handling of deletes received in a partial topk that
bring it below k records, by doing the same thing we do for extremum and
just querying the parent for the remaining records.

The rationale for why this is ok to do without evicting downstream is
the same as for extremum - if we receive an update, then
we're *guaranteed* that that update didn't exist in our parent, so
there's *no way* we'd ever miss when making a query into that parent.

This comes with a logictest that covers the delete issue in both the
partial and fully materialized cases, and a vertical test covering the
partial case.

Fixes: ENG-56
Change-Id: Ib27e64c26dd495f2714008188521f23d5c581e62
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/728
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>