vertical-tests: Add Update operation
Allow generating Update operations in the vertical testing suite, which
when executed issue an UPDATE query to the base table filtering on that
table's primary key. To generate them, we look at all previously
generated rows and generate a new row for the table where each column
has a 50/50 chance of being equal to the previous value for that column.
No-op updates (where no columns are changed) are filtered out.

Maybe unsurprisingly, the current query we're testing vertical tests
with seems to pass completely with this change! It failed once while I
was developing before I added a WHERE clause to the update, and I've
committed the seed for that failure, but otherwise things seem to be
happy.

Fixes: ENG-666
Change-Id: I76d1752ea47d13078246ce85deff46a1980d8842
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/726
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>