Cache some schema components for prepared statements
Currently noria-mysql performs a lot of repeated computations
when executing a prepared statement. Namely noria to mysql
schema conversion and column mapping. Some of those is actually
performed per resultset + per row, which leads to a large overhead.
With this change we will now cache the precomputed results of those
mappings.

Change-Id: Ida47951dc23dc09d907786fd2c5b9bcae8b15d56
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/735
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI