ci: Only run vertical tests in nightly
Ignore the test functions we generate with vertical tests by default
unless the new `vertical_tests` cargo feature is enabled (ignore, not
disable entirely, so we still check everything compiles on regular
branch builds). Then, add a new step to the nightly pipeline to run the
vertical test suite with this feature flags enabled, and the
PROPTEST_CASES env var set to 1000.

At this point, this is mostly happening because these tests are pretty
slow, not because they're uncovering any bugs (yet).

Fixes: ENG-669
Change-Id: Ia4e20362443230237fcd5d4b4fd88552cad58827
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/732
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI